% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsapy-dependencies.R
\name{add_gsapy_deps}
\alias{add_gsapy_deps}
\title{gsapy dependencies utils}
\usage{
add_gsapy_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches gsapy dependencies to the given tag
}
