% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsapy.R
\name{updateGsapy}
\alias{updateGsapy}
\title{Update GSAP animation}
\usage{
updateGsapy(id, animation, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The id of the GSAP animation to update.}

\item{animation}{The new animation to apply. Currently supported are
\itemize{
\item "fadeIn": Fade in elements that appear in the viewport on scroll, and
fade out elements that disappear from the viewport.
\item "zoomIn": Zoom in (scale) elements that appear in the viewport on scroll,
and zoom out elements that disappear from the viewport.
}}
}
\description{
You can use this function from the server side to change the
animation effect, for example from fadeIn to zoomIn.
}
